+++ MonomParse/IExpressionParser.cs
@@ -9,7 +9,7 @@ namespace MonomialParse
     public interface IExpressionParser
     {
         int? ExtractExponent(string expression);
-        decimal ExtractCoefficient(string expression);
+        int ExtractCoefficient(string expression);
         string ExtractVariable(string expression);
         string CombineStringExpression(decimal? coefficient, string variable, int? exponent);
